BackEnd/authentication_service/requirements.txt
Flask==0.10.1
mysql-connector
